THOR-1316 Moved the enums out of the abstract search service and fixed circular dependency errors.
