+++ src/app/components/text-cloud/text-cloud.component.ts
@@ -24,13 +24,7 @@ import {
     ViewEncapsulation
 } from '@angular/core';
 
-import {
-    AbstractSearchService,
-    CompoundFilterType,
-    FilterClause,
-    QueryPayload,
-    SortOrder
-} from '../../services/abstract.search.service';
+import { AbstractSearchService, FilterClause, QueryPayload } from '../../services/abstract.search.service';
 import { InjectableColorThemeService } from '../../services/injectable.color-theme.service';
 import { DashboardService } from '../../services/dashboard.service';
 import { FilterCollection, FilterDesign, SimpleFilterDesign } from '../../util/filter.util';
@@ -40,7 +34,9 @@ import { BaseNeonComponent } from '../base-neon-component/base-neon.component';
 import { NeonFieldMetaData } from '../../models/dataset';
 import {
     AggregationType,
+    CompoundFilterType,
     OptionChoices,
+    SortOrder,
     WidgetFieldOption,
     WidgetOption,
     WidgetSelectOption
